requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
fpdf2
